huggingface/transformers/sentence-similarity/sentence-t5-xxl/requirements.txt
optimum-rbln
huggingface_hub==0.30.2
sentence-transformers==3.4.1
